SAST Thunder monster
https://ygorganization.com/sast-cards-from-the-commerical/

Aware of how the thunder monster is not confirmed, but thought it would be best to prepare ahead of time. If it is indeed in Savage Strike, we can apply any necessary changes to update to match with the revealed effects. If it not confirmed and nowhere to be seen in the list of cards for Savage Strike, we will just simply remove the script.