c100235026
Not completed, but most of the structure is laid out (first effect is covered mostly, second effect still needs to be worked on in terms of spop). Thought it would be best leaving it here, as I must leave for an unknown period of time while having no access to my workspace. I thank you in advance, whoever willingly decides to continue on this script.